move user to features/share/func
